Placement positioning
Implemented a system for centering the monsters on the opposite side of the map and grouped by teams (clustered positioning).
Implemented a system for spacing the monsters on the map. This can be toggled on and off.
Implemented a system for layout of the monsters on the map. This can be toggled between clustered, columns, and random.